Fix segmentation fault from not initializing cache
Forgetting to call CACHE_INIT() in uszram_init() caused both cur0 and
cur1 of all caches to be zero at the beginning. This violated
assumptions made by the list2-cache functions, causing an unsigned
variable in uncache_pg() used as the "count" parameter of memmove to
underflow, which sometimes, unpredictably, caused segmentation faults,
seemingly only with multiple threads and when I wasn't looking
(single-stepping with gdb). Very hard to debug.